HystrixCollapser 使用请求合并器

+++ study-03-eureka-client/src/main/java/top/dpdaidai/cloud/study03eurekaclient/entity/User.java
@@ -7,7 +7,7 @@ package top.dpdaidai.cloud.study03eurekaclient.entity;
  */
 public class User {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -16,16 +16,16 @@ public class User {
 
     }
 
-    public User(Integer id, String name) {
+    public User(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
